Make It Easier To Start With Automation
This change reduces the effort needed for new users to start with automation and also improves the saving and loading of script files.
The new navigation in the Bot section introduces shortcuts to bots which are well suited for beginners and provides additional guidance on automation in eve online.
With this change, the IDE loads the script from the file which was last saved or loaded into the script editor on startup.
The UI for saving and loading script files is changed to work more like in common popular software for improved familiarity for new users.