Improve Accumulation Of Tooltips For Ship Modules
Adapt to observation in live tests: Ignore tooltips which contain no text. This assumes that every tooltip contains at least one `LabelText`. In case that does not apply anymore, we need to change the filter here to reflect this.
Thanks to pikacuq for sharing observations at https://forum.botengine.org/t/pikacuq-is-tweaking-a-bot/954/95?u=viir

Besides the accumulation filter, also change the casing of locals to match more popular convention.