Simplify mouse effects API
Make the computation of the region to move the mouse to transparent: Remove dependencies outside the part of the code which is easily observed by the bot developers. In cases were the 'relative to original' part is desired, it is better kept in the bot code since this is better accessible.

Also, to quote pikacuq:
> 2] work with relative grid to left upper corner of eve window. Which would be easier to measure in any graphical program so everybody would know what region he is working in based on exact coordinates.
This would be handy to specify exact areas of interest no matter which element is currently there. Of course there is a downside with counting them correctly when window size is changed.

For more information about this design, see the discussion at https://forum.botengine.org/t/there-is-a-way-to-mouse-click-on-entry-without-clicking-on-info-button/1301