Support development of memory measurement parsing
Fix a bottleneck in EVE Online bot development: Make the memory measurement parsing code easier accessible for bot developers: Support adapting the parsing code without dependency on Visual Studio and .NET build. To support the development of the parsing code, make it easier to collect sample inputs for the parsing functions on the bot code side: Add an example of how to save serialized versions of the different memory measurement stages to files.

In the past, adapting the memory measurement parsing code required a developer to set up Visual Studio and use a .NET build. To remove this friction, the parsing code is moved into the bot scope, where it can be changed as easily as your own bot code.
This commit adds an example of how to write the serialized memory measurements to files, so these can be easily imported when experimenting with changes to the parsing code