Explore memory reading from 64-bit EVE Online client
The tool added in this commit reads the UI tree from the 64-bit variant of the EVE Online client. It supports saving a process sample to a file for later inspection. The process samples contain all memory contents, so we can test memory reading implementations without the need to run a game client.

The `read-memory-eve-online` command can use a live process or a sample file as the source.

When using the `--output-file` argument, the tool writes the EVE Online UI tree into a JSON file.

From the looks of the JSON representation, we probably still find everything needed for in-game activities like ratting and mission running. Maybe we need to expand on some of the dictionary entries (e.g., background color) to get all the details we want, but there is no indication we get less than from the 32-bit client today.