Harden the memory reading implementation
Adapt to a sporadically observed problem in memory reading: On 2020-02-25 and 2020-03-04 observed issue caused the whole memory reading to fail with a `System.OverflowException` out of `ReadBytes`, in both cases called via `ReadActiveDictionaryEntriesFromDictionaryAddress`.
See the added image for an example of the error.

Improve the dictionary reading:
+ Avoiding invoking `ReadBytes` with an invalid argument
+ Avoid stalling the whole reading process when we happen to read garbage in a single dictionary.

For the report from 2020-03-04, see the discussion with McMahon on Discord.