Support reading python object type "Link"
Today we discovered the old way of reading 'Current Solar System' did not work anymore. It turned out that the '_setText' property of the label now references an object of type 'Link', which in turn contained a reference to a dictionary. That dictionary contains the properties of interest, three strings with keys '_alt', '_text', and '_url'. The '_text' property contained the current solar system name string as shown to the user.